feat(ui/assertion-ux):  move new routes constant to global.ts file
