+++ datahub-web-react/src/app/ProtectedRoutes.tsx
@@ -4,25 +4,19 @@ import { Layout } from 'antd';
 import { HomePage } from './home/HomePage';
 import { SearchRoutes } from './SearchRoutes';
 import EmbedRoutes from './EmbedRoutes';
-import { PageRoutes } from '../conf/Global';
-import { getRedirectUrl } from './shared/utils';
+import { NEW_ROUTE_MAP, PageRoutes } from '../conf/Global';
+import { getRedirectUrl } from '../conf/utils';
 
 /**
  * Container for all views behind an authentication wall.
  */
-export const ProtectedRoutes = (): JSX.Element => {
+export const ProtectedRoutes = (): JSX.Element => { 
     const location = useLocation();
     const history = useHistory();
 
     useEffect(() => {
         if (location.pathname.indexOf('/Validation') !== -1) {
-            const newRoutes = {
-                '/Validation/Assertions': '/Quality/List',
-                '/Validation/Tests': '/Governance/Tests',
-                '/Validation/Data%20Contract': '/Quality/Data%20Contract',
-                '/Validation': '/Quality',
-            };
-            history.replace(getRedirectUrl(newRoutes));
+            history.replace(getRedirectUrl(NEW_ROUTE_MAP));
         }
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [location]);
